refactor view stores to use newly created plots tab store
