+++ src/pages/studyView/StudyViewPlotsTabWrapper.tsx
@@ -13,60 +13,74 @@ export const PlotsTabWrapper: React.FunctionComponent<{
             filteredSamplesByDetailedCancerType={
                 store.filteredSamplesByDetailedCancerType
             }
-            mutations={store.mutations}
-            studies={store.queriedPhysicalStudies}
+            mutations={store.plotsTabStore.mutations}
+            studies={store.plotsTabStore.studies}
             molecularProfileIdSuffixToMolecularProfiles={
-                store.molecularProfileIdSuffixToMolecularProfiles
+                store.plotsTabStore.molecularProfileIdSuffixToMolecularProfiles
             }
-            entrezGeneIdToGene={store.entrezGeneIdToGeneAll}
-            sampleKeyToSample={store.sampleSetByKey}
-            genes={store.allGenes}
-            clinicalAttributes={store.clinicalAttributes}
-            genesets={store.genesets}
+            entrezGeneIdToGene={store.plotsTabStore.entrezGeneIdToGene}
+            sampleKeyToSample={store.plotsTabStore.sampleKeyToSample}
+            genes={store.plotsTabStore.genes}
+            clinicalAttributes={store.plotsTabStore.clinicalAttributes}
+            genesets={store.plotsTabStore.genesets}
             genericAssayEntitiesGroupByMolecularProfileId={
-                store.genericAssayEntitiesGroupedByProfileId
+                store.plotsTabStore
+                    .genericAssayEntitiesGroupByMolecularProfileId
             }
-            customAttributes={store.customAttributes}
-            studyIds={store.queriedPhysicalStudyIds}
-            molecularProfilesWithData={store.molecularProfilesInStudies}
-            molecularProfilesInStudies={store.molecularProfilesInStudies}
-            annotatedCnaCache={store.annotatedCnaCache}
-            annotatedMutationCache={store.annotatedMutationCache}
-            structuralVariantCache={store.structuralVariantCache}
+            customAttributes={store.plotsTabStore.customAttributes}
+            studyIds={store.plotsTabStore.studyIds}
+            molecularProfilesWithData={
+                store.plotsTabStore.molecularProfilesWithData
+            }
+            molecularProfilesInStudies={
+                store.plotsTabStore.molecularProfilesInStudies
+            }
+            annotatedCnaCache={store.plotsTabStore.annotatedCnaCache}
+            annotatedMutationCache={store.plotsTabStore.annotatedMutationCache}
+            structuralVariantCache={store.plotsTabStore.structuralVariantCache}
             studyToMutationMolecularProfile={
-                store.studyToMutationMolecularProfile
+                store.plotsTabStore.studyToMutationMolecularProfile
             }
             studyToMolecularProfileDiscreteCna={
-                store.studyToMolecularProfileDiscreteCna
+                store.plotsTabStore.studyToMolecularProfileDiscreteCna
             }
-            clinicalDataCache={store.clinicalDataCache}
-            patientKeyToFilteredSamples={store.patientKeyToFilteredSamples}
-            numericGeneMolecularDataCache={store.numericGeneMolecularDataCache}
-            coverageInformation={store.coverageInformation}
-            filteredSamples={store.selectedSamples}
-            genesetMolecularDataCache={store.genesetMolecularDataCache}
+            clinicalDataCache={store.plotsTabStore.clinicalDataCache}
+            patientKeyToFilteredSamples={
+                store.plotsTabStore.patientKeyToFilteredSamples
+            }
+            numericGeneMolecularDataCache={
+                store.plotsTabStore.numericGeneMolecularDataCache
+            }
+            coverageInformation={store.plotsTabStore.coverageInformation}
+            filteredSamples={store.plotsTabStore.filteredSamples}
+            genesetMolecularDataCache={
+                store.plotsTabStore.genesetMolecularDataCache
+            }
             genericAssayMolecularDataCache={
-                store.genericAssayMolecularDataCache
+                store.plotsTabStore.genericAssayMolecularDataCache
             }
             studyToStructuralVariantMolecularProfile={
-                store.studyToStructuralVariantMolecularProfile
+                store.plotsTabStore.studyToStructuralVariantMolecularProfile
             }
-            driverAnnotationSettings={store.driverAnnotationSettings}
-            studyIdToStudy={store.studyIdToStudy.result}
-            structuralVariants={store.structuralVariants.result}
-            hugoGeneSymbols={store.allHugoGeneSymbols.result}
+            driverAnnotationSettings={
+                store.plotsTabStore.driverAnnotationSettings
+            }
+            studyIdToStudy={store.plotsTabStore.studyIdToStudy}
+            structuralVariants={store.plotsTabStore.structuralVariants.result}
+            hugoGeneSymbols={store.plotsTabStore.hugoGeneSymbols}
             selectedGenericAssayEntitiesGroupByMolecularProfileId={
-                store.selectedGenericAssayEntitiesGroupByMolecularProfileId
+                store.plotsTabStore
+                    .selectedGenericAssayEntitiesGroupByMolecularProfileId
             }
             molecularProfileIdToMolecularProfile={
-                store.molecularProfileIdToMolecularProfile
+                store.plotsTabStore.molecularProfileIdToMolecularProfile
             }
             urlWrapper={urlWrapper}
-            hasNoQueriedGenes={true}
+            hasNoQueriedGenes={store.plotsTabStore.hasNoQueriedGenes}
             genePanelDataForAllProfiles={
-                store.genePanelDataForAllProfiles.result
+                store.plotsTabStore.genePanelDataForAllProfiles.result
             }
-            patients={store.patients}
+            patients={store.plotsTabStore.patients}
         />
     );
 });
