+++ typings/missing.d.ts
@@ -1,5 +1,6 @@
 // allow these file patterns to be imported
 declare module '*.scss';
+declare module '*.css';
 declare module '*.json';
 declare module '*.md';
 declare module '*.svg';
